# swagger.yaml
openapi: 3.0.3
info:
  title: Student Management System — API Gateway
  description: |
    Unified API Gateway for the University Student Management System.
    All requests are proxied to the appropriate microservice via the `/api` prefix.

    ### Microservice Routing
    | Prefix            | Target Service          |
    |-------------------|-------------------------|
    | `/api/students`   | Student & Auth Service  |
    | `/api/auth`       | Student & Auth Service  |
    | `/api/courses`    | Course Service          |
    | `/api/enrollments`| Enrollment Service      |
    | `/api/enroll`     | Enrollment Service      |
    | `/api/grades`     | Grade Service           |
    | `/api/gpa`        | Grade Service           |
  version: 1.0.0
  contact:
    name: Y4 Systems — SE4010
  license:
    name: MIT
servers:
  - url: https://api-gateway-763150334229.us-central1.run.app
    description: Production (Google Cloud Run)
  - url: http://localhost:8080
    description: Local Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /api/auth/login
  schemas:
...
    Student:
      type: object
      properties:
        _id:
          type: string
          example: "60f7b2c9e1d3a4001f3e7a01"
        name:
          type: string
          example: "Nuwan Fonseka"
        email:
          type: string
          example: "nuwan@university.edu"
        programme:
          type: string
          example: "BSc Computer Science"
        status:
          type: string
          enum: [Active, Inactive, Graduated]
          example: "Active"

    # ── Auth ────────────────────────────────
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          example: "admin@university.edu"
        password:
          type: string
          example: "password123"
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIs..."
        user:
          $ref: '#/components/schemas/Student'
    TokenGenerateRequest:
      type: object
      required: [userId, email, role]
      properties:
        userId:
          type: string
          example: "507f1f77bcf86cd799439011"
        email:
          type: string
          example: "student@university.edu"
        role:
          type: string
          enum: [admin, student]
          example: "student"
    TokenGenerateResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.sample.payload"
        expiresIn:
          type: string
          example: "24h"

    # ── Course ──────────────────────────────
    Course:
      type: object
      properties:
        _id:
          type: string
          example: "60f7b2c9e1d3a4001f3e7b02"
        name:
          type: string
          example: "Cloud Computing"
        courseId:
          type: string
          example: "SE4010"
        description:
          type: string
          example: "Introduction to cloud-native architectures"
        credits:
          type: integer
          example: 3

    # ── Enrollment ──────────────────────────
    Enrollment:
      type: object
      properties:
        _id:
          type: string
          example: "60f7b2c9e1d3a4001f3e7c03"
        student_id:
          type: string
          example: "S1001"
        course_id:
          type: string
          example: "C2002"
        status:
          type: string
          enum: [ACTIVE, CANCELLED, WITHDRAWN, COMPLETED]
          example: "ACTIVE"
        enrolled_at:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EnrollRequest:
      type: object
      required: [student_id, course_id]
      properties:
        student_id:
          type: string
          example: "S1001"
        course_id:
          type: string
          example: "C2002"

    # ── Grade ───────────────────────────────
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  example: "Nuwan Fonseka"
                email:
                  type: string
                  example: "nuwan@university.edu"
                password:
                  type: string
                  example: "password123"
      responses:
        '201':
          description: Student registered successfully
        '400':
          description: Validation error
        '409':
          description: Email already exists
  /api/auth/validate:
    get:
      tags: [Auth]
      summary: Validate JWT token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '401':
          description: Invalid or expired token
  /api/token/generate:
    post:
      tags: [Auth]
      summary: Generate JWT token for testing
      description: Generates a test JWT token with role `admin` or `student`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenGenerateRequest'
            examples:
              adminTokenRequest:
                summary: Generate admin token
                value:
                  userId: "507f1f77bcf86cd799439001"
                  email: "admin@university.edu"
                  role: "admin"
              studentTokenRequest:
                summary: Generate student token
                value:
                  userId: "507f1f77bcf86cd799439011"
                  email: "student@university.edu"
                  role: "student"
      responses:
        '200':
          description: Token generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenGenerateResponse'
              examples:
                adminTokenResponse:
                  summary: Admin token response
                  value:
                    token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.admin.payload"
                    expiresIn: "24h"
                studentTokenResponse:
                  summary: Student token response
                  value:
                    token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.student.payload"
                    expiresIn: "24h"
        '400':
          description: Invalid input or unsupported role

  # ── Students ────────────────────────────────
...
  /api/courses:
    get:
      tags: [Courses]
      summary: Get all courses
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'

    post:
      tags: [Courses]
      summary: Create a new course
      description: Requires JWT with role `admin`.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, courseId]
              properties:
                name:
                  type: string
                  example: "Cloud Computing"
                courseId:
                  type: string
                  example: "SE4010"
                description:
                  type: string
                credits:
                  type: integer
                  example: 3
      responses:
        '201':
          description: Course created
        '400':
          description: Validation error
...
  /api/enroll:
    post:
      tags: [Enrollments]
      summary: Enroll a student in a course
      description: |
        Validates student and course via external services, creates enrollment, and triggers Grade Service to create a grade record.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollRequest'
      responses:
        '201':
          description: Enrollment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Enrollment created successfully"
                  enrollment:
                    $ref: '#/components/schemas/Enrollment'
        '400':
          description: Missing student_id or course_id
        '401':
          description: Unauthorized
        '409':
          description: Student already enrolled in this course
        '503':
          description: External microservice unreachable
...
  /api/enrollments:
    get:
      tags: [Enrollments]
      summary: Get all enrollments
      description: Returns a list of all enrollment records in the system (proxied to Enrollment Service).
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all enrollments
        '404':
          description: No enrollments found
        '500':
          description: Internal server error
...
  /api/grades:
    post:
      tags: [Grades]
      summary: Create a grade record
      description: Usually triggered automatically by Enrollment Service after enrollment. Requires JWT with role `admin`.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [student_id, course_id]
              properties:
                student_id:
                  type: string
                  example: "S1001"
                course_id:
                  type: string
                  example: "C2002"
      responses:
        '201':
          description: Grade record created
...